# docs/contracts/gateway-openapi.yaml
openapi: 3.0.3
info:
  title: WikiAI Gateway API
  version: 0.1.0
  description: Public, admin, health, and external API contract for WikiAI Gateway.
servers:
  - url: http://127.0.0.1:3000
paths:
...
  /api/admin/reindex:
    post:
      summary: Start Syncer reindex from the admin plane.
      security:
        - mediaWikiSession: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReindexRequest"
      responses:
        "202":
          description: Reindex job accepted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/reindex/status:
    get:
      summary: Return latest Syncer reindex status.
      security:
        - mediaWikiSession: []
      responses:
        "200":
          description: Reindex status.
...
components:
  securitySchemes:
    mediaWikiSession:
      type: apiKey
      in: cookie
      name: mediawiki_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    internalAdminToken:
      type: apiKey
      in: header
      name: x-wikiai-admin-token
  responses:
...
    Unauthorized:
      description: Authentication is missing or invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Caller is not allowed to use the endpoint.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ReindexRequest:
      type: object
      properties:
        profileId:
          type: string
        indexTargets:
          type: array
          items:
            type: string
            enum: [dense, bm25, colbert, opensearch, attachments, semanticFacts, ontologyVectors]
        source:
          type: string
          enum: [mediawiki, qdrant_payload]
        maxPages:
          type: integer
          minimum: 1
        dryRun:
          type: boolean
        cmdbDynamicPagesEnabled:
          type: boolean
          description: Per-run override for cmdbdynamicpages marker processing.
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string